# Update `hexSlice` to use TypedArrays and TextDecoder for large speed gains